legacy/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
azure-identity
